feat: split action button into component
