--- client/src/app/(metadraft)/(validator)/data-validation/assets/actions.tsx
@@ -1,180 +1,0 @@
-import { useRouter } from "next/navigation";
-import { useRxCollection } from "rxdb-hooks";
-import { Button } from "~/components/ui/button";
-import { Typography } from "~/components/typography";
-import ArrowRightIcon from "~/icons/arrow-right.icon";
-import CheckIcon from "~/icons/check.icon";
-import FlagIcon from "~/icons/flag.icon";
-import type {
-  MetadataCollection,
-  ProjectCollection,
-  Status,
-  ValidationsCollection,
-} from "~/lib/types";
-import { cn } from "~/lib/utils";
-import { useActiveProject } from "~/providers/active-project.provider";
-import { keys } from "~/lib/constant";
-import TrashIcon from "~/icons/trash.icon";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "~/components/ui/dialog";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "~/components/ui/tooltip";
-
-export default function Actions({
-  metadata,
-  className,
-}: {
-  metadata: MetadataCollection;
-  className?: string;
-}) {
-  const activeProject = useActiveProject();
-  const router = useRouter();
-  const projectCollection = useRxCollection<ProjectCollection>("project");
-  const metadataCollection = useRxCollection<MetadataCollection>("metadata");
-  const validationsCollection =
-    useRxCollection<ValidationsCollection>("validations");
-
-  const project = activeProject?.toJSON() as ProjectCollection;
-
-  if (!metadata || !project) return <div>No metadata found</div>;
-
-  const isUnchecked = metadata.status === "unchecked";
-  const isSuccess = metadata.status === "success";
-  const isWarning = metadata.status === "warning";
-  const isError = metadata.status === "error";
-
-  const handleStatusUpdate = async (state: Status) => {
-    const currentState = metadata.status;
-    await metadataCollection?.upsert({
-      ...metadata,
-      status: state === currentState ? "error" : state,
-    });
-
-    const newProject = {
-      ...project,
-    };
-    // Remove stats
-    newProject[keys[currentState]]--;
-    // Add stats
-    newProject[keys[state === currentState ? "error" : state]]++;
-
-    // Add project information in RXDB
-    await projectCollection?.upsert(newProject);
-  };
-
-  // todo - add menu for card smaller UI ?
-
-  const handleDelete = async () => {
-    await metadataCollection?.bulkRemove([metadata.id]);
-
-    await projectCollection?.upsert({
-      ...project,
-      nfts: project.nfts - 1,
-      valids: project.valids - (isSuccess ? 1 : 0),
-      errorsDetected: project.errorsDetected - (isError ? 1 : 0),
-      errorsFlagged: project.errorsFlagged - (isWarning ? 1 : 0),
-      unchecked: project.unchecked - (isUnchecked ? 1 : 0),
-    });
-
-    await validationsCollection?.bulkRemove([metadata.id]);
-  };
-
-  return (
-    <div className={cn("flex flex-row items-center gap-2", className)}>
-      <TooltipProvider>
-        <Tooltip delayDuration={200}>
-          <TooltipTrigger asChild>
-            <Button
-              title="Mark as flagged"
-              disabled={isUnchecked}
-              variant={isWarning ? "warning" : "warningOutilne"}
-              size={"icon"}
-              onClick={async () => {
-                await handleStatusUpdate("warning");
-              }}
-            >
-              <FlagIcon className="h-4 w-4" />
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent className="max-w-97 flex flex-col gap-2 text-balance border-white/20 p-4 text-center">
-            <Typography as="smallText">{`Mark this asset as error-flagged.`}</Typography>
-            <Typography as="smallText">{`If it is already flagged, it will be marked as an error.`}</Typography>
-          </TooltipContent>
-        </Tooltip>
-      </TooltipProvider>
-
-      <TooltipProvider>
-        <Tooltip delayDuration={200}>
-          <TooltipTrigger asChild>
-            <Button
-              title="Mark as valid"
-              disabled={isUnchecked}
-              variant={isSuccess ? "success" : "successOutline"}
-              size={"icon"}
-              onClick={async () => {
-                await handleStatusUpdate("success");
-              }}
-            >
-              <CheckIcon className="h-4 w-4" />
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent className="max-w-97 flex flex-col gap-2 text-balance border-white/20 p-4 text-center">
-            <Typography as="smallText">{`Mark this asset as valid.`}</Typography>
-            <Typography as="smallText">{`If it is already valid, it will be marked as an error.`}</Typography>
-          </TooltipContent>
-        </Tooltip>
-      </TooltipProvider>
-
-      <Dialog>
-        <DialogTrigger asChild>
-          <Button
-            title="Delete asset"
-            size={"icon"}
-            variant={"destructiveOutilne"}
-          >
-            <TrashIcon className="h-4 w-4" />
-          </Button>
-        </DialogTrigger>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Are you absolutely sure?</DialogTitle>
-            <DialogDescription className="text-red-500">
-              This action will permanently remove the asset from your metadata
-              list. Once deleted, it cannot be undone. Please confirm if you
-              wish to proceed
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button
-              variant={"secondary"}
-              onClick={handleDelete}
-              className="w-full"
-            >
-              Confirm
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-      <Button
-        title="Detail asset"
-        size={"icon"}
-        variant={"outline"}
-        className="border-white/50"
-        onClick={() => router.push(`/metadata/${metadata.id}`)}
-      >
-        <ArrowRightIcon className="h-4 w-4" />
-      </Button>
-    </div>
-  );
-}
